+++ src/main/java/com/kt/ai/dto/response/OpenAIResponse.java
@@ -1,4 +1,4 @@
-package com.kt.ai.dto;
+package com.kt.ai.dto.response;
 
 import java.util.List;
